refactor: relocate safeToUseInsomniaFileName functions to sync/git/insomnia-filename
