+++ packages/insomnia/src/ui/components/modals/new-workspace-modal.tsx
@@ -22,7 +22,7 @@ import { useFetcher, useParams } from 'react-router';
 
 import { isGitProject, type Project } from '../../../models/project';
 import { type WorkspaceScope, WorkspaceScopeKeys } from '../../../models/workspace';
-import { safeToUseInsomniaFileName, safeToUseInsomniaFileNameWithExt } from '../../routes/actions';
+import { safeToUseInsomniaFileName, safeToUseInsomniaFileNameWithExt } from '../../../sync/git/insomnia-filename';
 import type { GetRepositoryDirectoryTreeResult } from '../../routes/git-project-actions';
 import type { StorageRules } from '../../routes/organization';
 import { Icon } from '../icon';
